Core/Entities: Changed object lowguid to uint64

+++ sql/base/auth_database.sql
@@ -434,7 +434,7 @@ DROP TABLE IF EXISTS `logs_ip_actions`;
 CREATE TABLE `logs_ip_actions` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT 'Unique Identifier',
   `account_id` int(10) unsigned NOT NULL COMMENT 'Account ID',
-  `character_guid` int(10) unsigned NOT NULL COMMENT 'Character Guid',
+  `character_guid` bigint(20) unsigned NOT NULL COMMENT 'Character Guid',
   `type` tinyint(3) unsigned NOT NULL,
   `ip` varchar(15) NOT NULL DEFAULT '127.0.0.1',
   `systemnote` text COMMENT 'Notes inserted by system',
